cleanup list management routes and add sentry

+++ app/jobs/unsubscribe_from_list_job.rb
@@ -5,13 +5,9 @@ class UnsubscribeFromListJob < ApplicationJob
     @client = Pardot.client
     subscription = Subscription.find(subscription_id)
 
-    prospect_id = salesforce_to_prospect(subscription.contact.salesforce_id)
+    return unless prospect_id = Pardot.salesforce_to_prospect(subscription.contact.salesforce_id)
 
     @client.list_memberships.delete(subscription.list.pardot_id, prospect_id)
     subscription.destroy
-  end
-
-  def salesforce_to_prospect(salesforce_id)
-    @client.prospects.read_by_fid(salesforce_id)['id']
   end
 end
